backEnd/requirements.txt
fastapi
uvicorn 
sqlalchemy 
pymysql
bcrypt
python-dotenv 
pydantic
jwt